Add files via upload
A Javascript app that uses the SpeechSynthesis interface of the Web Speech API. It's the controller interface for the speech service. this can be used to retrieve information about the synthesis voices available on your device, with commands like play, pause, and adjust the speed for a much faster or slower output. I added a little CSS to make it look like your are writing on paper.